Moved up kernel to -8MB of the address space, added mcmodel=tiny for aarch64 and made optimizations O3

+++ src/include/arch/x86/vmm.hpp
@@ -19,38 +19,38 @@ namespace x86_64 {
 
 #define PAGE_SIZE 0x1000
 
-#define VMM_FLAGS_ACCESSIBLE (1 << x86_64::PT_Flag::Present)
-#define VMM_FLAGS_NOEXEC     (1 << x86_64::PT_Flag::NX)
-#define VMM_FLAGS_WRITE      (1 << x86_64::PT_Flag::ReadWrite)
-#define VMM_FLAGS_NOCACHE    (1 << x86_64::PT_Flag::CacheDisabled)
-#define VMM_FLAGS_USER       (1 << x86_64::PT_Flag::UserSuper)
+#define VMM_FLAGS_ACCESSIBLE ((usize)1 << x86_64::PT_Flag::Present)
+#define VMM_FLAGS_NOEXEC     ((usize)1 << x86_64::PT_Flag::NX)
+#define VMM_FLAGS_WRITE      ((usize)1 << x86_64::PT_Flag::ReadWrite)
+#define VMM_FLAGS_NOCACHE    ((usize)1 << x86_64::PT_Flag::CacheDisabled)
+#define VMM_FLAGS_USER       ((usize)1 << x86_64::PT_Flag::UserSuper)
 
 #define VMM_FLAGS_KERNEL_CODE (VMM_FLAGS_ACCESSIBLE)
 
 #define VMM_FLAGS_KERNEL_RODATA (VMM_FLAGS_ACCESSIBLE)
 
-#define VMM_FLAGS_KERNEL_DATA (1 << x86_64::PT_Flag::Present | \
-			       1 << x86_64::PT_Flag::NX | \
-			       1 << x86_64::PT_Flag::ReadWrite)
+#define VMM_FLAGS_KERNEL_DATA ((usize)1 << x86_64::PT_Flag::Present | \
+			       (usize)1 << x86_64::PT_Flag::NX | \
+			      (usize)1 << x86_64::PT_Flag::ReadWrite)
 
-#define VMM_FLAGS_KERNEL_GENERIC (1 << x86_64::PT_Flag::Present | \
-			          1 << x86_64::PT_Flag::ReadWrite)
+#define VMM_FLAGS_KERNEL_GENERIC ((usize)1 << x86_64::PT_Flag::Present | \
+			         (usize)1 << x86_64::PT_Flag::ReadWrite)
 
 
-#define VMM_FLAGS_USER_CODE (1 << x86_64::PT_Flag::Present | \
-			     1 << x86_64::PT_Flag::UserSuper)
+#define VMM_FLAGS_USER_CODE ((usize)1 << x86_64::PT_Flag::Present | \
+			    (usize)1 << x86_64::PT_Flag::UserSuper)
 
-#define VMM_FLAGS_USER_RODATA (1 << x86_64::PT_Flag::Present | \
-			       1 << x86_64::PT_Flag::NX | \
-			       1 <<x86_64::PT_Flag::UserSuper)
+#define VMM_FLAGS_USER_RODATA ((usize)1 << x86_64::PT_Flag::Present | \
+			      (usize)1 << x86_64::PT_Flag::NX | \
+			      (usize)1 <<x86_64::PT_Flag::UserSuper)
 
-#define VMM_FLAGS_USER_DATA (1 << x86_64::PT_Flag::Present | \
-			     1 << x86_64::PT_Flag::NX | \
-			       1 << x86_64::PT_Flag::ReadWrite | \
-			       1 << x86_64::PT_Flag::UserSuper)
+#define VMM_FLAGS_USER_DATA ((usize)1 << x86_64::PT_Flag::Present | \
+			    (usize)1 << x86_64::PT_Flag::NX | \
+			      (usize)1 << x86_64::PT_Flag::ReadWrite | \
+			      (usize)1 << x86_64::PT_Flag::UserSuper)
 
-#define VMM_FLAGS_USER_GENERIC (1 << x86_64::PT_Flag::Present | \
-		                1 << x86_64::PT_Flag::ReadWrite | \
+#define VMM_FLAGS_USER_GENERIC ((usize)1 << x86_64::PT_Flag::Present | \
+		               (usize)1 << x86_64::PT_Flag::ReadWrite | \
 				1 << x86_64::PT_Flag::UserSuper)
 
 #define PTE_ADDR_MASK 0x000ffffffffff000
